third commit future_calculator.py

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
 monthly_income = int(input("Enter your monthly income: "))
 monthly_expenses = int(input("Enter your total monthly expenses: "))
-monthly_savings = int(monthly_income - monthly_expenses)
+monthly_savings = float(monthly_income) - float(monthly_expenses)
 ann_sav = int(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
